Fix how DUP downloader is initialized
The fix done in eef856c for the newer version of rclone (1.58.1) is
incomplete as our current tests show. This fix changes the downloader to
use the config full URL as its srcURL. Also when initilizing a
new HTTPFs, NewFs() returns an error and the fs object in case the given URL is
 a file (which is the case for DUP files), so we don't bubble up the error
in this case.