internal: Refactor different firmware detection
The current way we determine if there is a change in the firmware data,
is by comparing individual fields. This is not ideal as changes or
additional fileds can break this check.

This change is intended to make the check more maintainable by simply
marshaling the data to JSON and doing a bytes compare, which will allow
us to compare all fields even if the struct is updated.